requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
av==16.1.0
bidict==0.23.1
boto3==1.42.44
botocore==1.42.44
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
clip @ git+https://github.com/ultralytics/CLIP.git@88ade288431a46233f1556d1e141901b3ef0a36b
cloudpickle==3.1.2
coloredlogs==15.0.1
contourpy==1.3.2
ctranslate2==4.7.1
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
docutils==0.22.4
exceptiongroup==1.3.1
Farama-Notifications==0.0.4
fastapi==0.128.3
faster-whisper==1.2.1
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
ftfy==6.3.1
gymnasium==1.2.3
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.4.1
humanfriendly==10.0
idna==3.11
ifaddr==0.2.0
isaacsim==4.5.0.0
isaacsim-app==4.5.0.0
isaacsim-asset==4.5.0.0
isaacsim-benchmark==4.5.0.0
isaacsim-code-editor==4.5.0.0
isaacsim-core==4.5.0.0
isaacsim-cortex==4.5.0.0
isaacsim-example==4.5.0.0
isaacsim-extscache-kit==4.5.0.0
isaacsim-extscache-kit-sdk==4.5.0.0
isaacsim-extscache-physics==4.5.0.0
isaacsim-gui==4.5.0.0
isaacsim-kernel==4.5.0.0
isaacsim-replicator==4.5.0.0
isaacsim-rl==4.5.0.0
isaacsim-robot==4.5.0.0
isaacsim-robot-motion==4.5.0.0
isaacsim-robot-setup==4.5.0.0
isaacsim-ros1==4.5.0.0
isaacsim-ros2==4.5.0.0
isaacsim-sensor==4.5.0.0
isaacsim-storage==4.5.0.0
isaacsim-template==4.5.0.0
isaacsim-test==4.5.0.0
isaacsim-utils==4.5.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
kiwisolver==1.4.9
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
omniverse-kit==106.5.0.162521
onnxruntime==1.23.2
opencv-python==4.11.0.86
orjson==3.11.7
osqp==0.6.7.post3
packaging==26.0
pillow==12.1.0
polars==1.38.1
polars-runtime-32==1.38.1
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
PyAudio==0.2.14
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pyperclip==1.11.0
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.1
python-multipart==0.0.22
python-socketio==5.16.1
PyYAML==6.0.3
pyzmq==27.1.0
qdldl==0.1.7.post5
regex==2026.1.15
requests==2.32.5
rich==14.3.2
s3transfer==0.16.0
scipy==1.15.3
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sounddevice==0.5.5
SpeechRecognition==3.14.5
starlette==0.52.1
sympy==1.14.0
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
triton==3.6.0
typer==0.23.1
typer-slim==0.23.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.14
ultralytics-thop==2.0.18
urllib3==2.6.3
uv==0.10.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
wsproto==1.3.2
yarl==1.22.0
